Refactor(ajsf/core): doc, imports, methods
