+++ projects/ajsf-core/src/lib/framework-library/no-framework.component.html
@@ -1,0 +1,2 @@
+<select-widget-widget [dataIndex]="dataIndex" [layoutIndex]="layoutIndex" [layoutNode]="layoutNode">
+</select-widget-widget>
