--- projects/ajsf-core/src/lib/framework-library/no-framework/public_api.ts
@@ -1,3 +1,0 @@
-export * from './no.framework';
-export * from './no-framework.module';
-export * from './no-framework.component';
