+++ projects/ajsf-core/src/lib/framework-library/no.framework.ts
@@ -1,5 +1,5 @@
-import { Framework } from '../framework';
 import { Injectable } from '@angular/core';
+import { Framework } from './framework';
 import { NoFrameworkComponent } from './no-framework.component';
 // No framework - plain HTML controls (styles from form layout only)
