--- projects/ajsf-core/src/lib/framework-library/public_api.ts
@@ -1,5 +1,0 @@
-export { FrameworkLibraryService } from './framework-library.service';
-export { Framework } from "./framework";
-
-export { NoFramework } from './no-framework/no.framework';
-export { NoFrameworkModule } from './no-framework/no-framework.module';
