+++ projects/ajsf-core/src/lib/json-schema-form.module.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { NgModule } from '@angular/core';
 import { JsonSchemaFormComponent } from './json-schema-form.component';
-import { NoFrameworkModule } from './framework-library/public_api';
+import { NoFrameworkModule } from './framework-library/no-framework.module';
 import { WidgetLibraryModule } from './widget-library/widget-library.module';
 
 @NgModule({
