+++ projects/ajsf-core/src/lib/locale/index.ts
@@ -1,0 +1,4 @@
+export { enValidationMessages } from './en-validation-messages';
+export { frValidationMessages } from './fr-validation-messages';
+export { zhValidationMessages } from './zh-validation-messages';
+export { itValidationMessages } from './it-validation-messages';
