+++ projects/ajsf-core/src/lib/widget-library/widget-library.module.ts
@@ -1,22 +1,14 @@
 import { BASIC_WIDGETS } from './index';
 import { CommonModule } from '@angular/common';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { JsonSchemaFormService } from '../json-schema-form.service';
-import { ModuleWithProviders, NgModule } from '@angular/core';
+import { NgModule } from '@angular/core';
 import { OrderableDirective } from './orderable.directive';
 
 @NgModule({
-  imports:         [ CommonModule, FormsModule, ReactiveFormsModule ],
-  declarations:    [ ...BASIC_WIDGETS, OrderableDirective ],
-  exports:         [ ...BASIC_WIDGETS, OrderableDirective ],
-  entryComponents: [ ...BASIC_WIDGETS ],
-  providers:       [ JsonSchemaFormService ]
+  imports: [CommonModule, FormsModule, ReactiveFormsModule],
+  declarations: [...BASIC_WIDGETS, OrderableDirective],
+  exports: [...BASIC_WIDGETS, OrderableDirective],
+  entryComponents: [...BASIC_WIDGETS]
 })
 export class WidgetLibraryModule {
-  static forRoot(): ModuleWithProviders {
-    return {
-      ngModule: WidgetLibraryModule,
-      providers: [ JsonSchemaFormService ]
-    };
-  }
 }
