+++ projects/ajsf-material/src/lib/widgets/flex-layout-root.component.html
@@ -1,0 +1,11 @@
+<div *ngFor="let layoutNode of layout; let i = index" [class.form-flex-item]="isFlexItem"
+  [style.flex-grow]="getFlexAttribute(layoutNode, 'flex-grow')"
+  [style.flex-shrink]="getFlexAttribute(layoutNode, 'flex-shrink')"
+  [style.flex-basis]="getFlexAttribute(layoutNode, 'flex-basis')"
+  [style.align-self]="(layoutNode?.options || {})['align-self']" [style.order]="layoutNode?.options?.order"
+  [fxFlex]="layoutNode?.options?.fxFlex" [fxFlexOrder]="layoutNode?.options?.fxFlexOrder"
+  [fxFlexOffset]="layoutNode?.options?.fxFlexOffset" [fxFlexAlign]="layoutNode?.options?.fxFlexAlign">
+  <select-framework-widget *ngIf="showWidget(layoutNode)"
+    [dataIndex]="layoutNode?.arrayItem ? (dataIndex || []).concat(i) : (dataIndex || [])"
+    [layoutIndex]="(layoutIndex || []).concat(i)" [layoutNode]="layoutNode"></select-framework-widget>
+  <div>
